requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

dnspython>=1.14.0;python_version=='2.7'  # http://www.dnspython.org/LICENSE
eventlet>=0.17.4  # MIT
greenlet>=0.3.1
netifaces>=0.5,!=0.10.0,!=0.10.1
PasteDeploy>=1.3.3
lxml
requests>=2.14.2 # Apache-2.0
six>=1.9.0
xattr>=0.4
PyECLib>=1.3.1                          # BSD
cryptography!=2.0,>=1.6                 # BSD/Apache-2.0
# grpcio will fail to work with eventlet starting with 1.3.5.
# see this for a similar issue with gevent: https://github.com/grpc/grpc/issues/4629 and https://github.com/gevent/gevent/issues/786
# don't use eventlet for the object-server ? 
grpcio==1.3.3
fusepy>=2.0.4
